refactor: Moved the `supports_color` function to a more relevant location
