+++ src/logging.py
@@ -1,5 +1,4 @@
 import logging
-import os
 import re
 import sys
 from logging.handlers import RotatingFileHandler
@@ -7,6 +6,8 @@ from pathlib import Path
 
 import colorama
 from colorama import Fore, Style
+
+from src.utils.terminal import supports_color
 
 
 class ColorFormatter(logging.Formatter):
@@ -135,47 +136,6 @@ class Logger(logging.Logger):
         if self.isEnabledFor(self.SUCCESS):
             self._log(self.SUCCESS, msg, args, **kwargs)
 
-    @staticmethod
-    def supports_color() -> bool:
-        """Check if the terminal supports ANSI color codes.
-
-        Detects if the terminal supports ANSI color codes by checking platform-specific
-        conditions and environment variables. On Windows, it also checks the Windows
-        registry for the VirtualTerminalLevel key.
-
-        Returns:
-            bool: True if the terminal supports color, False otherwise
-        """
-
-        def vt_codes_enabled_in_windows_registry():
-            try:
-                import winreg
-            except ImportError:
-                return False
-
-            try:
-                reg_key = winreg.OpenKey(winreg.HKEY_CURRENT_USER, "Console")
-                reg_key_value, _ = winreg.QueryValueEx(reg_key, "VirtualTerminalLevel")
-                return reg_key_value == 1
-            except FileNotFoundError:
-                return False
-
-        is_a_tty = hasattr(sys.stdout, "isatty") and sys.stdout.isatty()
-
-        if not is_a_tty:
-            return False
-
-        if sys.platform == "win32":
-            return (
-                getattr(colorama, "fixed_windows_console", False)
-                or "ANSICON" in os.environ
-                or "WT_SESSION" in os.environ  # Windows Terminal
-                or os.environ.get("TERM_PROGRAM") == "vscode"
-                or vt_codes_enabled_in_windows_registry()
-            )
-
-        return True
-
     def setup(self, log_level: str, log_dir: str) -> None:
         """Configure the logger with console and file output.
 
@@ -187,7 +147,7 @@ class Logger(logging.Logger):
             log_dir (str): Directory where log files will be stored
         """
         has_color_support = False
-        if self.supports_color():
+        if supports_color():
             try:
                 if sys.platform == "win32":
                     colorama.just_fix_windows_console()
